fix: wrap useSearchParams in Suspense boundary on login page
Next.js requires useSearchParams() to be inside a Suspense boundary
for static generation. Extracted LoginForm into a separate client
component wrapped in Suspense.
